Added line to delete columns A and B in new files
Columns A and B contain data for the macro to create tab delimited text files for each compound object. Added a line in the macro to delete these two columns as they aren't typically in CONTENTdm metadata. A comment is in the macro if you would like to delete the line and keep those two columns.